Add getting user transactions by jwt access token

+++ prisma/migrations/20250323223223_init/migration.sql
@@ -16,7 +16,7 @@ CREATE TABLE "Product" (
 CREATE TABLE "Transaction" (
     "id" TEXT NOT NULL,
     "user_id" TEXT NOT NULL,
-    "amount" INTEGER NOT NULL,
+    "amount" DOUBLE PRECISION NOT NULL,
     "action" TEXT NOT NULL,
     "product_id" TEXT NOT NULL,
     "created_at" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
